Recreate DB file to accomodate changes in schema

+++ app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
-from app.db import Post, create_db_and_table, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from app.images import imagekit
 from app.users import auth_backend, current_active_user, fastapi_users
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -16,7 +16,7 @@ import tempfile
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     """Automatically creating database immediately after the app starts."""
-    await create_db_and_table()
+    await create_db_and_tables()
     yield
 
 
@@ -55,6 +55,7 @@ app.include_router(
 async def upload_file(
     file: UploadFile = File(...),
     caption: str = Form(""),  # Not request body, but request form
+    user: User = Depends(current_active_user),
     session: AsyncSession = Depends(
         get_async_session
     ),  # Dependency injection, calling get_async_session each time this endpoint is hit
@@ -77,6 +78,7 @@ async def upload_file(
             )
         if upload_result.response_metadata.http_status_code == 200:
             post = Post(
+                user_id=user.id,
                 caption=caption,
                 url=upload_result.url,
                 file_type="video" if file.content_type.startswith("video") else "image",
@@ -98,26 +100,37 @@ async def upload_file(
 
 
 @app.get("/feed")
-async def get_feed(session: AsyncSession = Depends(get_async_session)):
+async def get_feed(
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user),
+):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+    
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
                 "craeted_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user.id == post.user_id
             }
         )
     return {"posts": posts_data}
 
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(
+    post_id: str,
+    session: AsyncSession = Depends(get_async_session),
+    user: User = Depends(current_active_user),
+):
     try:
         post_uuid = uuid.UUID(post_id)
         result = await session.execute(select(Post).where(Post.id == post_uuid))
@@ -125,6 +138,11 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(
+                status_code=403, details="You don't have permission to delete this post"
+            )
 
         await session.delete(post)
         await session.commit()
